fixed html page for traffic lights monitoring and agents protocol

+++ src/acts/agents/state/traffic_light_state.py
@@ -13,7 +13,7 @@ class LightStatus:
 
 @dataclass
 class TrafficLightRuntimeState:
-    queue_length : int = 0
+    queue_length : float = 0
     waiting_time : int = 0
     status : LightStatus = LightStatus.RED
     status_time : int = 0   # For how long the light has been in the current status
